+++ service-layer/governanceApi/governanceApi/src/main/java/com/eric/governanceApi/governanceApi/repository/FirmwareRepository.java
@@ -1,7 +1,8 @@
 package com.eric.governanceApi.governanceApi.repository;
 
 import com.eric.governanceApi.governanceApi.enums.FirmwareStatus;
-import com.eric.governanceApi.governanceApi.model.Firmware;
+import com.eric.governanceApi.governanceApi.model.entity.Firmware;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
@@ -9,11 +10,11 @@ import java.util.Optional;
 
 public interface FirmwareRepository extends JpaRepository<Firmware, Long> {
 
-    Optional<Firmware> findByVersion(int version);
+    Optional<Firmware> findByVersion(float version);
 
     List<Firmware> findByStatusOrderByVersionDesc(FirmwareStatus status);
 
     List<Firmware> findAllByOrderByVersionDesc();
 
-    boolean existsByVersion(int version);
+    boolean existsByVersion(float version);
 }
